creditXpert/pyproject.toml
[project]
name = "CreditXpert"
version = "0.1.0"
description = "CreditXpert Flask App"
readme = "README.md"
authors = "Denis Burakov"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.0",
    "langchain-community>=0.3.18",
    "openai==0.28.1",
    "pydocstyle>=6.3.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "ruff>=0.9.9",
    "sentence-transformers>=3.4.1",
    "sqlite-vec>=0.1.6",
    "sqlite-vss>=0.1.2",
    "tomli>=2.2.1",
]